+++ choco-solver/src/main/java/solver/constraints/binary/PropGreaterOrEqualXY_C.java
@@ -54,7 +54,7 @@ public final class PropGreaterOrEqualXY_C extends Propagator<IntVar> {
 
     @SuppressWarnings({"unchecked"})
     public PropGreaterOrEqualXY_C(IntVar[] vars, int c) {
-        super(vars.clone(), PropagatorPriority.BINARY, true);
+        super(vars, PropagatorPriority.BINARY, true);
         this.x = vars[0];
         this.y = vars[1];
         this.cste = c;