+++ choco-solver/src/main/java/solver/constraints/binary/PropNotEqualXY_C.java
@@ -65,7 +65,7 @@ public class PropNotEqualXY_C extends Propagator<IntVar> {
 
     @SuppressWarnings({"unchecked"})
     public PropNotEqualXY_C(IntVar[] vars, int c) {
-        super(vars.clone(), PropagatorPriority.BINARY, true);
+        super(vars, PropagatorPriority.BINARY, true);
         this.x = vars[0];
         this.y = vars[1];
         this.cste = c;