+++ CHANGES.md
@@ -114,3 +114,4 @@ NEXT MILESTONE
 - Fix bug in CoupleTable due to wrong range use
 - Issue #191: disable buildFakeHistory by default (add a condition to build fake history)
 - Space are not filtered anymore from Operator
+- Remove vars.clone() in Propagator constructor.